+++ components/zcash_address/CHANGELOG.md
@@ -6,6 +6,8 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2022-10-19
 ### Added
 - `zcash_address::ConversionError`
 - `zcash_address::TryFromAddress`